+++ domain/src/test/kotlin/fr/sacane/jmanager/domain/fake/FakeFactory.kt
@@ -89,7 +89,8 @@ object FakeFactory {
         inMemoryRegularTransactionRepository,
         inMemoryTagRepository,
         sessionManager,
-        manager
+        manager,
+        inMemoryTrackerRepository
     )
     val fileImportExportFeature = FileImportExportFeatureImpl(
         csvFileReader,